+++ rules/standard/src/main/kotlin/dev/s7a/sqldelight/check/rules/standard/StandardRuleSetProvider.kt
@@ -1,20 +1,26 @@
 package dev.s7a.sqldelight.check.rules.standard
 
 import dev.s7a.sqldelight.check.api.RuleSetId
+import dev.s7a.sqldelight.check.rules.standard.rules.DataTypeCaseRule
 import dev.s7a.sqldelight.check.rules.standard.rules.FinalNewlineRule
+import dev.s7a.sqldelight.check.rules.standard.rules.FunctionNameCaseRule
 import dev.s7a.sqldelight.check.rules.standard.rules.KeywordCaseRule
 import dev.s7a.sqldelight.check.rules.standard.rules.LineEndingLfRule
+import dev.s7a.sqldelight.check.rules.standard.rules.LiteralCaseRule
 import dev.s7a.sqldelight.check.rules.standard.rules.MaxBlankLinesRule
 import dev.s7a.sqldelight.check.rules.standard.rules.NoConsecutiveSemicolonsRule
 import dev.s7a.sqldelight.check.rules.standard.rules.NoLeadingBlankLinesRule
+import dev.s7a.sqldelight.check.rules.standard.rules.NoSpaceAfterDotRule
 import dev.s7a.sqldelight.check.rules.standard.rules.NoSpaceAfterOpeningParenthesisRule
 import dev.s7a.sqldelight.check.rules.standard.rules.NoSpaceBeforeClosingParenthesisRule
 import dev.s7a.sqldelight.check.rules.standard.rules.NoSpaceBeforeCommaRule
+import dev.s7a.sqldelight.check.rules.standard.rules.NoSpaceBeforeDotRule
 import dev.s7a.sqldelight.check.rules.standard.rules.NoSpaceBeforeSemicolonRule
 import dev.s7a.sqldelight.check.rules.standard.rules.NoTabIndentationRule
 import dev.s7a.sqldelight.check.rules.standard.rules.NoTrailingBlankLinesRule
 import dev.s7a.sqldelight.check.rules.standard.rules.NoTrailingWhitespaceRule
 import dev.s7a.sqldelight.check.rules.standard.rules.SpaceAfterCommaRule
+import dev.s7a.sqldelight.check.rules.standard.rules.SpaceAfterLineCommentMarkerRule
 import dev.s7a.sqldelight.check.rules.standard.rules.SpaceAroundComparisonOperatorsRule
 import dev.s7a.sqldelight.check.rule.api.RuleProvider
 import dev.s7a.sqldelight.check.rule.api.RuleSetProvider
@@ -31,20 +37,26 @@ public class StandardRuleSetProvider : RuleSetProvider {
      */
     override fun ruleProviders(): Set<RuleProvider> =
         setOf(
+            RuleProvider(::DataTypeCaseRule),
             RuleProvider(::FinalNewlineRule),
+            RuleProvider(::FunctionNameCaseRule),
+            RuleProvider(::KeywordCaseRule),
             RuleProvider(::LineEndingLfRule),
+            RuleProvider(::LiteralCaseRule),
             RuleProvider(::MaxBlankLinesRule),
             RuleProvider(::NoConsecutiveSemicolonsRule),
             RuleProvider(::NoLeadingBlankLinesRule),
+            RuleProvider(::NoSpaceAfterDotRule),
             RuleProvider(::NoSpaceAfterOpeningParenthesisRule),
             RuleProvider(::NoSpaceBeforeClosingParenthesisRule),
             RuleProvider(::NoSpaceBeforeCommaRule),
+            RuleProvider(::NoSpaceBeforeDotRule),
             RuleProvider(::NoSpaceBeforeSemicolonRule),
             RuleProvider(::NoTabIndentationRule),
             RuleProvider(::NoTrailingBlankLinesRule),
             RuleProvider(::NoTrailingWhitespaceRule),
-            RuleProvider(::KeywordCaseRule),
             RuleProvider(::SpaceAfterCommaRule),
+            RuleProvider(::SpaceAfterLineCommentMarkerRule),
             RuleProvider(::SpaceAroundComparisonOperatorsRule),
         )
 }