+++ rules/standard/src/main/kotlin/dev/s7a/sqldelight/check/rules/standard/rules/KeywordCaseRule.kt
@@ -91,7 +91,6 @@ public class KeywordCaseRule : Rule {
                 "like",
                 "limit",
                 "not",
-                "null",
                 "on",
                 "or",
                 "order",